*adjust player shoot sound pos
git-svn-id: https://svn.code.sf.net/p/quake2xp/code@1017 d83f5309-02f3-4a42-a5f4-12e6af65da40

+++ trunk/qcommon/cvar.c
@@ -70,7 +70,7 @@ float Cvar_VariableValue (char *var_name) {
 	return atof (var->string);
 }
 
-float Cvar_VariableInteger(char *var_name) {
+int Cvar_VariableInteger(char *var_name) {
 	cvar_t *var;
 
 	var = Cvar_FindVar(var_name);
@@ -305,6 +305,8 @@ void Cvar_ForceSetValue (char *var_name, float value) {
 		Com_sprintf (val, sizeof(val), "%f", value);
 	Cvar_ForceSet (var_name, val);
 }
+
+
 
 /*
 ============
